+++ CHANGELOG.rst
@@ -18,6 +18,8 @@ Changed
 
 - Update ``graphql`` dependency to 0.2.0.
 
+- Disable failover on replicasets where number of instances less than 3.
+
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 Added
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~